refactor(daemon): extract SSE transport into http/sse.ts
Strangler-fig slice 2, cluster 2. Moves the Server-Sent Events transport
helper — createSseResponse and its default heartbeat cadence
SSE_KEEPALIVE_INTERVAL_MS — out of server.ts into http/sse.ts, alongside
the other http/ transport helpers (response.ts, api-errors.ts).

Byte-identical move. Both names were exported from server.ts, so server.ts
now imports createSseResponse back for its route deps objects and
re-exports both names — the public export surface is unchanged. Scope is
the SSE transport only; createSseErrorPayload (an ApiError-shaped payload
builder that depends on createCompatApiError and belongs with
http/api-errors) is intentionally left in place for a later api-errors move.

Validation: tsc -p apps/daemon/tsconfig.json --noEmit clean; pnpm guard
clean; chat-run-sse-shapes characterization 4/4.
