+++ apps/daemon/src/server.ts
@@ -56,6 +56,8 @@ export {
   createCompatApiError,
   createCompatApiErrorResponse,
 } from './http/api-errors.js';
+import { createSseResponse } from './http/sse.js';
+export { createSseResponse, SSE_KEEPALIVE_INTERVAL_MS } from './http/sse.js';
 import {
   applyBakedPreviews,
   resolvePluginPreviewsDir,
@@ -1090,7 +1092,6 @@ const critiqueWarnedAdapters = new Set<string>();
 // can cascade an AbortController to the matching orchestrator invocation.
 // Created once per process; not persisted across daemon restarts.
 const critiqueRunRegistry = createRunRegistry();
-export const SSE_KEEPALIVE_INTERVAL_MS = 25_000;
 
 export function createAgentRuntimeEnv(
   baseEnv: NodeJS.ProcessEnv | Record<string, string | undefined>,
@@ -3187,65 +3188,6 @@ function sendMulterError(res, err) {
   }
 
   return sendApiError(res, 500, 'INTERNAL_ERROR', 'upload failed');
-}
-
-export function createSseResponse(
-  res,
-  { keepAliveIntervalMs = SSE_KEEPALIVE_INTERVAL_MS } = {},
-) {
-  res.setHeader('Content-Type', 'text/event-stream');
-  res.setHeader('Cache-Control', 'no-cache, no-transform');
-  res.setHeader('Connection', 'keep-alive');
-  res.setHeader('X-Accel-Buffering', 'no');
-  res.flushHeaders?.();
-
-  const canWrite = () => !res.destroyed && !res.writableEnded;
-  const writeKeepAlive = () => {
-    if (canWrite()) {
-      res.write(': keepalive\n\n');
-      return true;
-    }
-    return false;
-  };
-
-  let heartbeat = null;
-  if (keepAliveIntervalMs > 0) {
-    heartbeat = setInterval(writeKeepAlive, keepAliveIntervalMs);
-    heartbeat.unref?.();
-  }
-
-  const cleanup = () => {
-    if (heartbeat) {
-      clearInterval(heartbeat);
-      heartbeat = null;
-    }
-  };
-
-  res.on('close', cleanup);
-  res.on('finish', cleanup);
-
-  return {
-    /** @param {ChatSseEvent['event'] | ProxySseEvent['event'] | string} event */
-    send(event, data, id: string | number | null | undefined = null) {
-      if (!canWrite()) return false;
-      // Assemble the full SSE event into a single write so id/event/data land
-      // in one TCP chunk. Three separate writes would let `event: <type>` flush
-      // ahead of the `data:` payload, which produces partial events for
-      // consumers that read chunk-by-chunk (e.g. tests using a Response body
-      // reader with a substring marker).
-      const idLine = id !== null && id !== undefined ? `id: ${id}\n` : '';
-      res.write(`${idLine}event: ${event}\ndata: ${JSON.stringify(data)}\n\n`);
-      return true;
-    },
-    writeKeepAlive,
-    cleanup,
-    end() {
-      cleanup();
-      if (canWrite()) {
-        res.end();
-      }
-    },
-  };
 }
 
 export type DesktopPdfExporter = (input: DesktopExportPdfInput) => Promise<DesktopExportPdfResult>;
